Sensor read update
can now read moves from the updated sensor array. Move takes will be implemented in the future as follows:
a 64th bit will be sent out from the sensor arrays. this 64th bit will represent the piece to be taken. In game, a player will first move the desired piece to the take sensor, press space, and then move the piece that is going to be doing the taking to the old pieces square. the chess manager will then check to make sure that the move is legal.