Bill's edition
This is Bill's contribution for CSE 426. Inside Bill's folder, there are two subfolders, one for the microprocessor containing an older Arduino code for self implimented movement, another for the ESP32 which is operating with grbl.
The other folder has a windows application which was used to create an interface for moving the h-bot/corexy system to check for binding and stress along with serial communication testing.